requirements.txt
prefect>=3.0.0
scrapling[all]
clickhouse-connect
kafka-python
